requirements.txt
haystack-ai
docling
pinecone
openai
python-dotenv
maxapi
requests
